# services/web-bff/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: HooshiX Web BFF Public API
  version: 2.0.0
  description: Canonical browser-facing REST contract owned by Web BFF. All examples use synthetic .test identities and non-production placeholder credentials.
servers:
  - url: /
security:
  - BrowserSession: []
paths:
...
  /api/v1/identity/mfa:
    get:
      operationId: getIdentityMfaStatus
      summary: Read the authenticated user's MFA state
      security: [{ BrowserSession: [] }]
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      x-hooshix-consumer-example: { method: GET, path: /api/v1/identity/mfa }
      responses:
        '200': { description: Current MFA state., content: { application/json: { schema: { $ref: '#/components/schemas/MfaStatusResponse' } } } }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '401': { $ref: '#/components/responses/InvalidSession' }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }
...
  /api/v1/identity/mfa/totp/enrollment:
    post:
      operationId: startIdentityTotpEnrollment
      summary: Start initial TOTP enrollment or an authenticated replacement
      description: Returns the Base32 secret once for authenticator configuration. Replacing active TOTP requires a current MFA proof.
      security: [{ BrowserSession: [] }]
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/CsrfToken' }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/StartTotpEnrollmentRequest' } } }
      x-hooshix-consumer-example: { method: POST, path: /api/v1/identity/mfa/totp/enrollment, body: { currentProof: { type: TOTP, code: '123456' } } }
      responses:
        '200': { description: Pending TOTP enrollment., content: { application/json: { schema: { $ref: '#/components/schemas/TotpEnrollmentResponse' } } } }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '401': { $ref: '#/components/responses/InvalidSession' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }
  /api/v1/identity/mfa/totp/enrollment/confirm:
    post:
      operationId: confirmIdentityTotpEnrollment
      summary: Confirm a pending TOTP enrollment
      description: Recovery codes are returned exactly once and must be stored offline by the user.
      security: [{ BrowserSession: [] }]
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/CsrfToken' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ConfirmTotpEnrollmentRequest' } } } }
      x-hooshix-consumer-example: { method: POST, path: /api/v1/identity/mfa/totp/enrollment/confirm, body: { enrollmentChallenge: YWJjZGVmZ2hpamtsbW5vcHFyc3R1dnd4eXoxMjM0NTY, totpCode: '123456' } }
      responses:
        '200': { description: TOTP enabled and recovery codes issued once., content: { application/json: { schema: { $ref: '#/components/schemas/RecoveryCodesResponse' } } } }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '401': { $ref: '#/components/responses/InvalidSession' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }
  /api/v1/identity/mfa/totp:
    delete:
      operationId: disableIdentityTotp
      summary: Disable TOTP after a current MFA proof
      security: [{ BrowserSession: [] }]
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/CsrfToken' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/MfaProofRequest' } } } }
      x-hooshix-consumer-example: { method: DELETE, path: /api/v1/identity/mfa/totp, body: { type: TOTP, code: '123456' } }
      responses:
        '200': { description: TOTP disabled and browser security state rotated. }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '401': { $ref: '#/components/responses/InvalidSession' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }
  /api/v1/identity/mfa/recovery-codes/rotate:
    post:
      operationId: rotateIdentityMfaRecoveryCodes
      summary: Revoke and replace all unused MFA recovery codes
      description: The replacement codes are returned exactly once.
      security: [{ BrowserSession: [] }]
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/CsrfToken' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/MfaProofRequest' } } } }
      x-hooshix-consumer-example: { method: POST, path: /api/v1/identity/mfa/recovery-codes/rotate, body: { type: TOTP, code: '123456' } }
      responses:
        '200': { description: Replacement recovery codes issued once., content: { application/json: { schema: { $ref: '#/components/schemas/RecoveryCodesResponse' } } } }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '401': { $ref: '#/components/responses/InvalidSession' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }
...
components:
  securitySchemes:
    BrowserSession:
      type: apiKey
      in: cookie
      name: __Host-sajtech-session
      description: HttpOnly same-origin session. Its value is never exposed in this contract or browser storage.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Caller-generated UUIDv4 idempotency identity. Replays use the same value.
      schema:
        type: string
        format: uuid
        pattern: '^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$'
...
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      description: Current session-bound synchronizer token returned by a reviewed BFF response; do not persist it.
      schema: { type: string, minLength: 32, maxLength: 512 }
...
  schemas:
...
    MfaProofRequest:
      example: { type: TOTP, code: '123456' }
      discriminator: { propertyName: type }
      oneOf:
        - type: object
          additionalProperties: false
          required: [type, code]
          properties:
            type: { type: string, const: TOTP, example: TOTP }
            code: { type: string, pattern: '^[0-9]{6}$', writeOnly: true, example: '123456' }
        - type: object
          additionalProperties: false
          required: [type, code]
          properties:
            type: { type: string, const: RECOVERY_CODE, example: RECOVERY_CODE }
            code: { type: string, pattern: '^[A-Z2-7]{4}(?:-[A-Z2-7]{4}){3}$', writeOnly: true, example: AAAA-BBBB-CCCC-DDDD }
    StartTotpEnrollmentRequest:
      type: object
      additionalProperties: false
      example: { currentProof: { type: TOTP, code: '123456' } }
      properties:
        currentProof:
          anyOf:
            - { $ref: '#/components/schemas/MfaProofRequest' }
            - { type: 'null' }
    ConfirmTotpEnrollmentRequest:
      type: object
      additionalProperties: false
      required: [enrollmentChallenge, totpCode]
      example: { enrollmentChallenge: YWJjZGVmZ2hpamtsbW5vcHFyc3R1dnd4eXoxMjM0NTY, totpCode: '123456' }
      properties:
        enrollmentChallenge: { type: string, pattern: '^[A-Za-z0-9_-]{43}$', writeOnly: true, example: YWJjZGVmZ2hpamtsbW5vcHFyc3R1dnd4eXoxMjM0NTY }
        totpCode: { type: string, pattern: '^[0-9]{6}$', writeOnly: true, example: '123456' }
    MfaStatusResponse:
      type: object
      additionalProperties: false
      required: [totpEnabled, recoveryCodesRemaining]
      example: { totpEnabled: true, recoveryCodesRemaining: 10 }
      properties:
        totpEnabled: { type: boolean, example: true }
        recoveryCodesRemaining: { type: integer, minimum: 0, maximum: 10, example: 10 }
    TotpEnrollmentResponse:
      type: object
      additionalProperties: false
      required: [enrollmentChallenge, base32Secret, otpauthUri, expiresAt]
      example: { enrollmentChallenge: YWJjZGVmZ2hpamtsbW5vcHFyc3R1dnd4eXoxMjM0NTY, base32Secret: AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA, otpauthUri: 'otpauth://totp/SajTech%3Asynthetic?secret=AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA&issuer=SajTech&algorithm=SHA256&digits=6&period=30', expiresAt: '2026-01-01T00:10:00Z' }
      properties:
        enrollmentChallenge: { type: string, pattern: '^[A-Za-z0-9_-]{43}$', readOnly: true }
        base32Secret: { type: string, pattern: '^[A-Z2-7]{52}$', readOnly: true }
        otpauthUri: { type: string, pattern: '^otpauth://totp/', maxLength: 512, readOnly: true }
        expiresAt: { type: string, format: date-time }
    RecoveryCodesResponse:
      type: object
      additionalProperties: false
      required: [recoveryCodes]
      example: { recoveryCodes: [AAAA-BBBB-CCCC-DDDD, EEEE-FFFF-GGGG-HHHH, IIII-JJJJ-KKKK-LLLL, MMMM-NNNN-OOOO-PPPP, QQQQ-RRRR-SSSS-TTTT, UUUU-VVVV-WWWW-XXXX, YYYY-ZZZZ-2222-3333, '4444-5555-6666-7777', ABCD-EFGH-IJKL-MNOP, QRST-UVWX-YZ23-4567] }
      properties:
        recoveryCodes:
          type: array
          minItems: 10
          maxItems: 10
          uniqueItems: true
          readOnly: true
          items: { type: string, pattern: '^[A-Z2-7]{4}(?:-[A-Z2-7]{4}){3}$' }
...
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status, code]
      example: { type: urn:hooshix:problem:invalid-request, title: Invalid request, status: 400, code: INVALID_REQUEST, instance: /api/v1/example }
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
          maxLength: 120
        status:
          type: integer
          minimum: 400
          maximum: 599
        code:
          type: string
          pattern: '^[A-Z][A-Z0-9_]{0,63}$'
        instance:
          type: string
          maxLength: 1024
  responses:
    InvalidRequest:
      description: Request validation failed.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    InvalidSession:
      description: A supplied browser session is invalid.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Forbidden:
      description: Browser-origin or session-bound request protection rejected the request.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    RateLimited:
      description: The applicable semantic security quota was exceeded.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    DependencyUnavailable:
      description: The BFF runtime or an authoritative downstream dependency is unavailable.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...